If a node is in multiple partitions, it would be listed multiple times in the sinfo output.
This would trigger an error for entry into the temporary database as a constraint issue.

By sorting the list of nodes initially, then piping that through the uniq command, we eliminate
duplicates.

When this script was first written, all nodes in the system were in one and only one partition,
thus this problem was never run across.